Only install vagrant-vbguest when explicitly asked
Using vagrant-vbguest adds around 10 minutes to the test duration, and
also opens of possibility of test failures due to bugs in the VirtualBox
Guest Additions (which will also taint the box kernel).

You can add an --with-vagrant-vbguest argument to install_vagrant.sh in
the cico configuration (cloudinstance-vagrant-test on ci.centos.org).